- deploy to kubernetes

+++ product-service/product-domain/product-application-service/src/main/java/com/anw/product/service/domain/dto/ProductBaseResponse.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -15,7 +16,7 @@ public class ProductBaseResponse {
     private final UUID id;
     private final String name;
     private final String description;
-    private final double price;
+    private final BigDecimal price;
     private final int stockQuantity;
     private final String imageUrl;
     private final CategoryResponse category;
